api-server/requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
gunicorn>=21.2.0
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cpu
transformers>=4.30.0
huggingface_hub>=0.16.0
sentencepiece
protobuf
